moved compete page
