+++ src/pages/CompetePage/leaderboard.css
